Implemented tester for Ex02
Tests 01 and 02 might work now.  Tests 03 and 04 are still stubs.

Also refactored _tester, moving most of it into test_lib.py, so that I could
reuse almost all of it in 02/_tester.

Implemented a differ for JSON that allows floats to disagree by 0.001.

Also hopefully fixed the differ for PNMs, so that it handles arbitrary spacing
in P1 type files.